+++ src/parser/block/table.test.ts
@@ -12,14 +12,10 @@ describe('Unit: parser/table', () => {
       assert.deepStrictEqual(inspect(parser('|')), void 0);
       assert.deepStrictEqual(inspect(parser('||')), void 0);
       assert.deepStrictEqual(inspect(parser('|h')), void 0);
-      assert.deepStrictEqual(inspect(parser('|h\n|-')), void 0);
       assert.deepStrictEqual(inspect(parser('|h|')), void 0);
       assert.deepStrictEqual(inspect(parser('|h|\n')), void 0);
       assert.deepStrictEqual(inspect(parser('|h|\n|')), void 0);
       assert.deepStrictEqual(inspect(parser('|h|\n||')), void 0);
-      assert.deepStrictEqual(inspect(parser('|h|\n|-')), void 0);
-      assert.deepStrictEqual(inspect(parser('|h|\n|-|')), void 0);
-      assert.deepStrictEqual(inspect(parser('|h|\n|-|\n')), void 0);
     });
 
     it('valid', () => {
@@ -36,13 +32,14 @@ describe('Unit: parser/table', () => {
       assert.deepStrictEqual(inspect(parser('|a|b|\n|-|-\n|1|2|')), [['<table><thead><tr><td align="">a</td><td align="">b</td></tr></thead><tbody><tr><td align="">1</td><td align="">2</td></tr></tbody></table>'], '']);
       assert.deepStrictEqual(inspect(parser('|a|b|\n|-|-\n|1|2')), [['<table><thead><tr><td align="">a</td><td align="">b</td></tr></thead><tbody><tr><td align="">1</td><td align="">2</td></tr></tbody></table>'], '']);
       assert.deepStrictEqual(inspect(parser('|a|b|\n|-|\n|1|2|')), [['<table><thead><tr><td align="">a</td><td align="">b</td></tr></thead><tbody><tr><td align="">1</td><td align="">2</td></tr></tbody></table>'], '']);
-      assert.deepStrictEqual(inspect(parser('|a|b|\n|-|\n|1|')), [['<table><thead><tr><td align="">a</td><td align="">b</td></tr></thead><tbody><tr><td align="">1</td></tr></tbody></table>'], '']);
-      assert.deepStrictEqual(inspect(parser('|a|b|\n|-\n|1')), [['<table><thead><tr><td align="">a</td><td align="">b</td></tr></thead><tbody><tr><td align="">1</td></tr></tbody></table>'], '']);
-      assert.deepStrictEqual(inspect(parser('|a|b|\n|-\n|')), [['<table><thead><tr><td align="">a</td><td align="">b</td></tr></thead><tbody><tr><td align=""></td></tr></tbody></table>'], '']);
-      assert.deepStrictEqual(inspect(parser('|a|b\n|-\n|')), [['<table><thead><tr><td align="">a</td><td align="">b</td></tr></thead><tbody><tr><td align=""></td></tr></tbody></table>'], '']);
-      assert.deepStrictEqual(inspect(parser('|a|\n|-\n|')), [['<table><thead><tr><td align="">a</td></tr></thead><tbody><tr><td align=""></td></tr></tbody></table>'], '']);
-      assert.deepStrictEqual(inspect(parser('|a\n|-\n|')), [['<table><thead><tr><td align="">a</td></tr></thead><tbody><tr><td align=""></td></tr></tbody></table>'], '']);
-      assert.deepStrictEqual(inspect(parser('|\n|-\n|')), [['<table><thead><tr><td align=""></td></tr></thead><tbody><tr><td align=""></td></tr></tbody></table>'], '']);
+      assert.deepStrictEqual(inspect(parser('|a|b|\n|-|\n|1|')), [['<table><thead><tr><td align="">a</td><td align="">b</td></tr></thead><tbody><tr><td align="">1</td><td align=""></td></tr></tbody></table>'], '']);
+      assert.deepStrictEqual(inspect(parser('|a|b|\n|-\n|1')), [['<table><thead><tr><td align="">a</td><td align="">b</td></tr></thead><tbody><tr><td align="">1</td><td align=""></td></tr></tbody></table>'], '']);
+      assert.deepStrictEqual(inspect(parser('|a|b|\n|-\n|')), [['<table><thead><tr><td align="">a</td><td align="">b</td></tr></thead><tbody><tr><td align=""></td><td align=""></td></tr></tbody></table>'], '']);
+      assert.deepStrictEqual(inspect(parser('|a|b|\n|-')), [['<table><thead><tr><td align="">a</td><td align="">b</td></tr></thead><tbody></tbody></table>'], '']);
+      assert.deepStrictEqual(inspect(parser('|a|b\n|-')), [['<table><thead><tr><td align="">a</td><td align="">b</td></tr></thead><tbody></tbody></table>'], '']);
+      assert.deepStrictEqual(inspect(parser('|a|\n|-')), [['<table><thead><tr><td align="">a</td></tr></thead><tbody></tbody></table>'], '']);
+      assert.deepStrictEqual(inspect(parser('|a\n|-')), [['<table><thead><tr><td align="">a</td></tr></thead><tbody></tbody></table>'], '']);
+      assert.deepStrictEqual(inspect(parser('|\n|-')), [['<table><thead><tr><td align=""></td></tr></thead><tbody></tbody></table>'], '']);
     });
 
     it('align', () => {